requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
django-extensions==3.2.3
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
